Reorganiza classes de domain models
