+++ app/artbook/domain/artist.py
@@ -44,3 +44,12 @@ class Artist():
         for altname in alternative_names:
             self.__alternative_names.discard(altname)
 
+
+class GroupOfArtists():
+    def __init__(self, name: str, foundation: date=None, end: date=None):
+        self.name = name
+        self.foundation = foundation
+        self.end = end
+
+    def __repr__(self):
+        return self.name
