+++ app/artbook/domain/artwork.py
@@ -24,3 +24,11 @@ class Artwork():
             'title': self.title,
             'creation': self.creation
         }
+
+
+class ArtworkSeries():
+    def __init__(self, name: str):
+        self.name = name
+
+    def __repr__(self):
+        return self.name
